requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pypdf==4.3.1
numpy==1.26.4
tqdm==4.66.5
python-multipart==0.0.9
pymupdf==1.24.9
pdfminer.six==20231228
